refactor: 불필요한 Lottos를 제거한 후 책임을 LottoService로 이동
